Fixed vertex cache, should double your framerate

+++ color.go
@@ -2,8 +2,8 @@ package sf
 
 // Represents an RGBA color with intensities between 0 and 255
 type Color struct {
-	R float32
-	G float32
-	B float32
-	A float32
+	R uint8
+	G uint8
+	B uint8
+	A uint8
 }
